% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findParetoSet}
\alias{findParetoSet}
\title{Find the pareto set, or the convex hull of the input data}
\usage{
findParetoSet(idealPoints)
}
\arguments{
\item{idealPoints}{...}
}
\description{
Find the pareto set, or the convex hull of the input data
}
