% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kintersection}
\alias{kintersection}
\title{Return k intersection}
\usage{
kintersection(polyList, k, epsilon)
}
\arguments{
\item{polyList}{An R list containing DataFrames with 2 columns names 'x' and 'y' representing polygon points}

\item{k}{Number of intersections user is looking for}

\item{epsilon}{Error tolerance}
}
\description{
Return k intersection
}
